content/html/content/test/test_bug595447.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=595447
-->
<head>
  <title>Test for Bug 595447</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=595447">Mozilla Bug 595447</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 595447 **/

var fieldset = document.createElement("fieldset");

ok("type" in fieldset, "fieldset element should have a type IDL attribute");
is(fieldset.type, "fieldset", "fieldset.type should return 'fieldset'");
fieldset.type = "foo";
is(fieldset.type, "fieldset", "fieldset.type is readonly");

</script>
</pre>
</body>
</html>
